Introduce Gl trait
So far we have basically only had an implicitly defined interface for
the lowest level GL functionality: both the OpenGL and the WebGL
backends were kept in-sync entirely manually.
As we extend the interface in ways where some functionality is only
added to one of the backends initially, this is problematic, because
divergence becomes just too easy.
With this change we fix this problem by introducing the Gl trait, which
is implement for both Context types, that explicitly defines the
interface. In so doing we get a (limited) sanity checking of both
interfaces even without the need to involve an actual consumer
application.